part2/hbnb/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-restx==1.3.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
urllib3==2.2.3
Werkzeug==3.1.2
